feat(worker): Tear out non-reuseV8Context support
