+++ packages/test/src/load/worker.ts
@@ -175,8 +175,6 @@ async function main() {
       interceptors: {
         activity: [() => ({ inbound: new ConnectionInjectorInterceptor(clientConnection) })],
       },
-      // Can't reuse the helper because it defines `test` and ava thinks it's an ava test.
-      reuseV8Context: ['1', 't', 'true'].includes((process.env.REUSE_V8_CONTEXT ?? 'false').toLowerCase()),
     });
 
     await withOptionalStatusServer(worker, statusPort, async () => {
